Add Client GUI Login, Register
The client part is using a basic MVC (Mode View Controller) stucture.
The Gui consists of a MainFrame in witch different panels are loaded in.

Loading the panels are determined by the main Controller, and other
event handler controllers (LoginController, RegisterController) whitch
are assigned to a single panel (LoginPanel, RegisterPanel).